Map gen

+++ core/src/Models/Entities/BlockModel.java
@@ -3,8 +3,8 @@ package Models.Entities;
 
 public class BlockModel extends EntityModel {
 
-    public BlockModel(int x, int y){
-        super(x,y);
+    public BlockModel(float x, float y){
+        super(x ,y);
     }
 
 }
